feat: update api

+++ larksuite-oapi/src/main/java/com/lark/oapi/service/hire/v1/model/AgencySupplierProtectTime.java
@@ -33,7 +33,7 @@ public class AgencySupplierProtectTime {
      * <p> 示例值：180
      */
     @SerializedName("day")
-    private Integer day;
+    private Long day;
     /**
      * 是否使用统一设置
      * <p> 示例值：true
@@ -62,11 +62,11 @@ public class AgencySupplierProtectTime {
         return new Builder();
     }
 
-    public Integer getDay() {
+    public Long getDay() {
         return this.day;
     }
 
-    public void setDay(Integer day) {
+    public void setDay(Long day) {
         this.day = day;
     }
 
@@ -83,7 +83,7 @@ public class AgencySupplierProtectTime {
          * 保护时长，单位（天）
          * <p> 示例值：180
          */
-        private Integer day;
+        private Long day;
         /**
          * 是否使用统一设置
          * <p> 示例值：true
@@ -97,7 +97,7 @@ public class AgencySupplierProtectTime {
          * @param day
          * @return
          */
-        public Builder day(Integer day) {
+        public Builder day(Long day) {
             this.day = day;
             return this;
         }
